+++ components/api_server/src/initialization/datamodel.py
@@ -1,23 +1,24 @@
 """Data model loader."""
 
 import json
-import logging
 
 from pymongo.database import Database
 
 from shared_data_model import DATA_MODEL_JSON
 
 from database.datamodels import insert_new_datamodel, latest_datamodel
+from utils.log import get_logger
 
 
 def import_datamodel(database: Database) -> None:  # pragma: no feature-test-cover
     """Store the data model in the database."""
+    logger = get_logger()
     data_model = json.loads(DATA_MODEL_JSON)
     if latest := latest_datamodel(database):
         del latest["timestamp"]
         del latest["_id"]
         if data_model == latest:
-            logging.info("Skipping loading the data model; it is unchanged")
+            logger.info("Skipping loading the data model; it is unchanged")
             return
     insert_new_datamodel(database, data_model)
-    logging.info("Data model loaded")
+    logger.info("Data model loaded")